backend/requirements.txt
Django==5.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-silk
setuptools
python-dateutil==2.8.2
psycopg2-binary==2.9.9
requests>=2.31.0